Remove use of RefersToAnyLN in LogControl.ocl
RefersToLog was not setup in LogControl, it is now.
RefersToAnyLN can be accessed with RefersToLog.ParentAnyLN, so it has been removed.
Moreover, if the Log, and therefore the AnyLN, have been found, it means that ldInst and prefix are correct, so it is useless to test in OCL that they are correct using RefersToAnyLN or RefersToLog! And if they are wrong, warning messages have already been displayed.